# backend/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "SchoolPay API",
    "version": "0.1.0-mvp",
    "description": "Release-readiness OpenAPI contract for SchoolPay auth, finance, notification, and support chat modules."
  },
  "servers": [
    {
      "url": "http://localhost:8080/api",
      "description": "Local API"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "PaymentRequest": {
        "type": "object",
        "required": [
          "student_id",
          "amount",
          "channel",
          "method"
        ],
        "properties": {
          "student_id": {
            "type": "integer",
            "minimum": 1
          },
          "amount": {
            "type": "number",
            "minimum": 1
          },
          "channel": {
            "type": "string",
            "example": "cash"
          },
          "method": {
            "type": "string",
            "example": "Tunai"
          },
          "reference": {
            "type": "string"
          },
          "bill_ids": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "is_bypass_rule": {
            "type": "boolean"
          }
        }
      },
      "PaymentIntentRequest": {
        "type": "object",
        "required": [
          "student_id",
          "amount",
          "bill_ids"
        ],
        "properties": {
          "student_id": {
            "type": "integer",
            "minimum": 1
          },
          "amount": {
            "type": "number",
            "minimum": 1
          },
          "bill_ids": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "is_bypass_rule": {
            "type": "boolean"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "paths": {
...
    "/finance/payments": {
      "post": {
        "summary": "Process manual/cash/deposit payment",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Processed"
          }
        }
      },
      "get": {
        "summary": "Get successful payment history by student_id",
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/finance/payment-intent": {
      "post": {
        "summary": "Create Midtrans payment intent/order",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PaymentIntentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Intent created"
          }
        }
      }
    },
...
  }
}